fix(migrate Popover): feedback
