+++ lib/src/components/Popover/Content.tsx
@@ -5,6 +5,8 @@ import { classNames } from '@/utils'
 import popoverStyles from './popover.module.scss'
 const cx = classNames(popoverStyles)
 
-export const Content = forwardRef<HTMLDivElement, React.HTMLProps<HTMLDivElement>>((props, ref) => {
-  return <div className={cx('content')} ref={ref} {...props} />
-})
+export const Content = forwardRef<HTMLDivElement, React.HTMLProps<HTMLDivElement>>(
+  ({ className, ...props }, ref) => {
+    return <div className={cx('content', className)} ref={ref} {...props} />
+  }
+)
