+++ lib/src/components/Popover/Title.tsx
@@ -6,7 +6,7 @@ import popoverStyles from './popover.module.scss'
 const cx = classNames(popoverStyles)
 
 export const Title = forwardRef<HTMLHeadingElement, React.HTMLProps<HTMLHeadingElement>>(
-  (props, ref) => {
-    return <h6 className={cx('title')} ref={ref} {...props} />
+  ({ className, ...props }, ref) => {
+    return <h6 className={cx('title', className)} ref={ref} {...props} />
   }
 )
