+++ lib/src/components/Popover/popover.module.scss
@@ -50,7 +50,7 @@
     color: var(--color-neutral-90);
   }
   .arrow-item {
-    transform: var(--transform);
+    transform: var(--popover-transform);
     width: 30px;
     height: 30px;
   }
