+++ website/build-app/pages/foundations/upgrades/v9.mdx
@@ -44,7 +44,7 @@ Size props are only for intern values (xs, sm, ...), you can now use tailwind si
 
 ```diff
 -<CheckIcon size="24"  />
-+<CheckIcon className="size-[24px]" />
++<CheckIcon className="size-(--spacing-xl)" />
 ```
 
 ### Popover
